src/api/requirements.txt

fastapi
uvicorn[standard]
python-dotenv
requests
python-multipart